+++ packages/components/ContextualMenu/CHANGELOG.md
@@ -1,8 +1,35 @@
 # Change Log - @fluentui-react-native/contextual-menu
 
-This log was last generated on Tue, 01 Mar 2022 20:18:31 GMT and should not be manually modified.
+This log was last generated on Fri, 04 Mar 2022 23:43:10 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.16.8
+
+Fri, 04 Mar 2022 23:43:10 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/callout to v0.19.30
+- Bump @fluentui-react-native/focus-zone to v0.9.7
+- Bump @fluentui-react-native/icon to v0.11.11
+- Bump @fluentui-react-native/interactive-hooks to v0.14.2
+- Bump @fluentui-react-native/pressable to v0.8.25
+
+## 0.16.7
+
+Thu, 03 Mar 2022 20:20:09 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/callout to v0.19.29
+- Bump @fluentui-react-native/focus-zone to v0.9.6
+- Bump @fluentui-react-native/icon to v0.11.10
+- Bump @fluentui-react-native/interactive-hooks to v0.14.1
+- Bump @fluentui-react-native/text to v0.11.12
+- Bump @fluentui-react-native/tokens to v0.11.7
+- Bump @uifabricshared/foundation-compose to v1.11.11
+- Bump @fluentui-react-native/pressable to v0.8.24
 
 ## 0.16.6