+++ packages/framework/framework/CHANGELOG.md
@@ -1,8 +1,26 @@
 # Change Log - @fluentui-react-native/framework
 
-This log was last generated on Fri, 18 Feb 2022 23:27:10 GMT and should not be manually modified.
+This log was last generated on Fri, 04 Mar 2022 23:43:10 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.7.16
+
+Fri, 04 Mar 2022 23:43:10 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/composition to v0.7.6
+
+## 0.7.15
+
+Thu, 03 Mar 2022 20:20:09 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/default-theme to v0.9.11
+- Bump @fluentui-react-native/theme-types to v0.14.0
+- Bump @fluentui-react-native/tokens to v0.11.7
 
 ## 0.7.14